LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-keyboard-events-expected.txt

Multiple fields UI of month input type with keyboard events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Please run this with DumpRenderTree.

Test following keys:
Digits
Left/Right - Move focus field inside element
Up/Down - Increment/decrement value of focus field
Tab - Move focus field
Backspace - Make value empty
  
== Digit keys ==
FAIL input.value should be 0012-09. Was 0112-09.
== Left/Right keys ==
FAIL input.value should be 0005-06. Was 0005-09.
PASS document.activeElement.id is "input"
== Up/Down keys ==
PASS input.value is "2012-10"
PASS input.value is "2012-08"
== Up/Down keys on empty value ==
PASS input.value is "2012-11"
== Up/Down keys on empty value 2 ==
PASS input.value is "2012-02"
== Tab key ==
PASS input.value is "0005-09"
PASS input.value is "0005-01"
PASS document.activeElement.id is "another"
== Shfit+Tab key ==
PASS input.value is "0003-09"
PASS document.activeElement.id is "before"
== Up key on maximum value ==
PASS input.value is "0001-10"
== Up key with a maximum attribute ==
PASS input.value is "0001-01"
== Down key on minimum value ==
PASS input.value is ""
== Down key with a minimum attribute ==
PASS input.value is ""
== Inconsistent min-max attributes ==
PASS input.value is "1000-12"
PASS input.value is "1999-12"
== Backspace key ==
PASS input.value is ""
== Delete key ==
PASS input.value is ""
== Typeahead ==
PASS input.value is "0001-12"
PASS input.value is "0002-12"
== RTL Left/Right keys ==
PASS input.value is "2012-10"
PASS input.value is "0002-10"
PASS input.value is "0002-11"
PASS successfullyParsed is true

TEST COMPLETE
